Nits and (c)
