+++ src/Microsoft.Extensions.Logging/ILoggerBuilder.cs
@@ -1,3 +1,6 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Microsoft.Extensions.Logging
@@ -8,15 +11,5 @@ namespace Microsoft.Extensions.Logging
         /// Gets the <see cref="IServiceCollection"/> where Logging services are configured.
         /// </summary>
         IServiceCollection Services { get; }
-    }
-
-    public class LoggerBuilder : ILoggerBuilder
-    {
-        public LoggerBuilder(IServiceCollection services)
-        {
-            Services = services;
-        }
-
-        public IServiceCollection Services { get; }
     }
 }
