--- src/Microsoft.Extensions.Logging/LogMessageFilter.cs
@@ -1,4 +1,0 @@
-namespace Microsoft.Extensions.Logging
-{
-    public delegate bool LogMessageFilter(string loggerType, string categoryName, LogLevel level);
-}
