+++ src/Microsoft.Extensions.Logging/LoggerBuilder.cs
@@ -1,14 +1,17 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.Extensions.Options;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Microsoft.Extensions.Logging
 {
-    public class DefaultLoggerLevelConfigureOptions : ConfigureOptions<LoggerFilterOptions>
+    internal class LoggerBuilder : ILoggerBuilder
     {
-        public DefaultLoggerLevelConfigureOptions(LogLevel level) : base(options => options.MinLevel = level)
+        public LoggerBuilder(IServiceCollection services)
         {
+            Services = services;
         }
+
+        public IServiceCollection Services { get; }
     }
 }
