+++ src/Microsoft.Extensions.Logging/LoggerFilterRule.cs
@@ -1,10 +1,13 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 
 namespace Microsoft.Extensions.Logging
 {
     public class LoggerFilterRule
     {
-        public LoggerFilterRule(string loggerType, string categoryName, LogLevel? logLevel, LogMessageFilter filter)
+        public LoggerFilterRule(string loggerType, string categoryName, LogLevel? logLevel, Func<string, string, LogLevel, bool> filter)
         {
             LoggerType = loggerType;
             CategoryName = categoryName;
@@ -18,6 +21,6 @@ namespace Microsoft.Extensions.Logging
 
         public LogLevel? LogLevel { get; }
 
-        public LogMessageFilter Filter { get; }
+        public Func<string, string, LogLevel, bool> Filter { get; }
     }
 }
