+++ src/Microsoft.Extensions.Logging/LoggerInformation.cs
@@ -1,0 +1,35 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+
+namespace Microsoft.Extensions.Logging
+{
+    internal struct LoggerInformation
+    {
+        public ILogger Logger { get; set; }
+
+        public string Category { get; set; }
+
+        public string ProviderType { get; set; }
+
+        public LogLevel? MinLevel { get; set; }
+
+        public Func<string, string, LogLevel, bool> Filter { get; set; }
+
+        public bool IsEnabled(LogLevel level)
+        {
+            if (MinLevel != null && level < MinLevel)
+            {
+                return false;
+            }
+
+            if (Filter != null)
+            {
+                return Filter(ProviderType, Category, level);
+            }
+
+            return true;
+        }
+    }
+}
